make changes at index_calculation
